rename and remove default rendering of ig
